Refactor leaderboard related code
